requirements.txt
fastapi
uvicorn
nltk
langchain
langchain_community
langchain_google_genai
faiss-cpu
pandas
matplotlib
seaborn
numpy
streamlit
google-generativeai
ipython
scikit-learn
scipy
unstructured
openpyxl
tiktoken
python-magic
networkx
starlette
jinja2
textwrap3
ipython
python-multipart
pillow
werkzeug
nest-asyncio
streamlit
wordcloud
nltk
spacy>=3.0.0,<4.0.0
PyMuPDF
python-dotenv
gensim
pyLDAvis